Aborting this mf
This project was made in 17-4-2025 when i was bored and i wanted to learn something. But since im retarded, lazy and i dont have nerves for this. I quickly gave up and now i dont want to work on this project. Maybe after some years i will get back to it